documents_web/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
Django==5.1.1
djangorestframework==3.15.2
minio==7.2.9
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
setuptools==74.1.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
wheel==0.37.1
